convert voter from feature to component, move files and update imports
