+++ src/app/store.js
@@ -3,7 +3,6 @@ import postModalReducer from '../features/PostModal/postModalSlice';
 import searchBarReducer from '../features/SearchBar/searchBarSlice';
 import subredditsReducer from '../features/Subreddits/subredditsSlice';
 import counterReducer from '../features/Counter/counterSlice';
-import voterReducer from '../features/Voter/voterSlice';
 import timerReducer from '../features/Timer/timerSlice';
 
 const store =  configureStore({
@@ -13,7 +12,6 @@ const store =  configureStore({
         subreddits: subredditsReducer,
         timer: timerReducer,
         counter: counterReducer,
-        voter: voterReducer
     }
 })
