+++ src/components/Voter/Voter.css
