+++ src/components/Voter/Voter.js
